// backend/src/types/error-keys.ts
export enum ErrorKeys {
  Invalid_credentials = 'invalid_credentials',
  Invalid_fields = 'invalid_fields',
  Invalid_password = 'invalid_password',
  Invalid_token = 'invalid_token',
  Missing_fields = 'invalid_credentials',
  Unauthorized = 'unauthorized',
  User_already_exists = 'user_already_exists',
  User_not_found = 'user_not_found',
  Validation_failed = 'validation_failed',
}